refactor: using generic types for connections and edges
